Introduce per-platform directories
This way we can have multiple implementations of platform-specific
functions instead of `#ifdef`s all over the place
