+++ userspace/libscap/linux/gettimeofday.h
@@ -1,0 +1,33 @@
+/*
+Copyright (C) 2022 The Falco Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
+
+#pragma once
+
+#include <sys/time.h>
+#include <stdint.h>
+
+static inline uint64_t get_timestamp_ns()
+{
+	uint64_t ts;
+	struct timeval tv;
+
+	gettimeofday(&tv, NULL);
+	ts = tv.tv_sec * (uint64_t) 1000000000 + tv.tv_usec * 1000;
+
+	return ts;
+}
+
